requirements.txt
django-crispy-forms
whitenoise
django-widget-tweaks
